Run CI test and lint on all branches
This makes pushes trigger CI checks even when they are not to the
main branch or in a PR opened against the main branch. This way,
CI will work in forks once enabled in the Actions tab (otherwise
the changes have to be on the "main" branch of the fork, or a PR
is needed).