Quote shell expansions
This is not really a bugfix, because it's possible to infer the
correctness of the unquoted expansions in the cases where they
appear in the CI workflows.

However, quoting them makes the intent clear: we want the literal
text produced by the expansion, and it is not intenteded that it be
changed by word splitting and globbing (filename expansion) done by
the shell.