Test on Python 3.12 experimentally
This allows a prerelease for 3.12, which is currently at RC2 so it
will not install if a prerelease is not allowed, while continuing
to prohibit installing a prerelease for earlier versions.

Because 3.12 is at RC2, the likelihood of test failures on it due
to bugs in 3.12 is fairly low. But in any situation where tests do
fail on 3.12, it would be necessary to compare the reults to those
of at least one other release, to rule this out. So I have also
set continue-on-error to true for 3.12 so that failure of a 3.12
job will not cancel the other jobs.