Test on three major operating systems
This extends the CI test matrix with an "os" dimension, so it tests
on Ubuntu (as before), macOS, and Windows. It tests all five Python
version on each, so the matrix now has 15 jobs.

Building the value for "os" by appending "-latest" to names like
"ubuntu", rather than just using the full names ("ubuntu-latest"),
is less idiomatic. However, it helps keep the generated matrix job
names narrow enough not to be cut off in the GitHub Actions
interface, for matrices like this one that have a third dimension.
(Currently, "core" is the only value for "test-type", but it is,
and probably should, be kept in the names, thus lengthening them.)