Install poetry with pip (not pipx) in test jobs
This is imperfect, but it allows the correct version of Python
to be installed and used with minimal complexity and special
casing. Furthermore, because poetry is being used and the project
is being installed in a virtual environment, polluting the global
Python environment with poetry itself and its dependencies is
unlikely to hide unstated dependency bugs or cause conflicts.

On Ubuntu runners, pipx works with whatever version of Python is
installed by setup-python, and commands like "python3.10" are
available in $PATH for pipx to find and use. That is not the case
on other GitHub Actions runners, however. This is especially a
problem on Windows, where even obtaining the full path to the
interpreter cannot be done in any straightforward portable way.
(setup-python provides the Python directory in the pythonLocation
Python_ROOT_DIR etc. environment variables, but where the Python
interpeter is in it can differ by OS. This is practical, though
cumbersome, to solve when only CPython is used. But if pypy is also
used, which may be desired at some point in the future, then even
more special-casing is required, on Windows.)